feat: move accounts pages to features/accounts module
Relocates all files from frontend/src/pages/accounts/ to
frontend/src/features/accounts/pages/ and updates all imports.

- Update features/accounts/index.ts to export AccountsPage and AccountDetailPage
- Update App.tsx to import from @/features/accounts barrel
- Update payment-form-utils.ts cross-reference to use @/features/accounts/pages/account-form-utils
