+++ frontend/src/App.tsx
@@ -13,8 +13,7 @@ import { AppShell } from '@/components/layout/app-shell'
 import { TooltipProvider } from '@/components/ui/tooltip'
 import { TenantsPage } from '@/pages/tenants/index'
 import { TenantDetailPage } from '@/pages/tenants/[tenantId]'
-import { AccountsPage } from '@/pages/accounts'
-import { AccountDetailPage } from '@/pages/accounts/[accountId]'
+import { AccountsPage, AccountDetailPage } from '@/features/accounts'
 import { PaymentsPage } from '@/pages/payments'
 import { PaymentDetailPage } from '@/pages/payments/payment-detail'
 import { PartiesPage } from '@/pages/parties'
