+++ frontend/src/features/accounts/index.ts
@@ -1,1 +1,3 @@
 // Public API for accounts module
+export { AccountsPage } from './pages/index'
+export { AccountDetailPage } from './pages/[accountId]'
