+++ frontend/src/pages/payments/dialogs/payment-form-utils.ts
@@ -3,4 +3,4 @@
  * Re-exports the string-based BigInt conversion from account-form-utils
  * to avoid IEEE-754 float precision issues in payment amount handling.
  */
-export { amountToBigInt, bigIntToDisplayAmount, formatCurrencyAmount } from '../../accounts/account-form-utils'
+export { amountToBigInt, bigIntToDisplayAmount, formatCurrencyAmount } from '@/features/accounts/pages/account-form-utils'
